# api/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: pgEdge Control Plane API
    description: Service for creating, modifying, and operating pgEdge databases.
    version: 0.0.1
host: localhost:3000
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /cluster:
        get:
            tags:
                - control-plane
            summary: inspect-cluster control-plane
            description: Returns information about the cluster.
            operationId: control-plane#inspect-cluster
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Cluster'
                        required:
                            - id
                            - tenant_id
                            - status
                            - hosts
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/ControlPlaneInspectClusterNotFoundResponseBody'
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ControlPlaneInspectClusterClusterNotInitializedResponseBody'
            schemes:
                - http
    /cluster/init:
        get:
            tags:
                - control-plane
            summary: init-cluster control-plane
            description: Initializes a new cluster.
            operationId: control-plane#init-cluster
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/ClusterJoinToken'
                        required:
                            - token
                            - server_url
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ControlPlaneInitClusterClusterAlreadyInitializedResponseBody'
            schemes:
                - http
    /cluster/join:
        post:
            tags:
                - control-plane
            summary: join-cluster control-plane
            description: Join this host to an existing cluster.
            operationId: control-plane#join-cluster
            parameters:
                - name: Join-ClusterRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/ClusterJoinToken'
                    required:
                        - token
                        - server_url
            responses:
                "204":
                    description: No Content response.
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ControlPlaneJoinClusterClusterAlreadyInitializedResponseBody'
            schemes:
                - http
    /cluster/join-token:
        get:
            tags:
                - control-plane
            summary: get-join-token control-plane
            description: Gets the join token for this cluster.
            operationId: control-plane#get-join-token
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/ClusterJoinToken'
                        required:
                            - token
                            - server_url
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ControlPlaneGetJoinTokenClusterNotInitializedResponseBody'
            schemes:
                - http
    /databases:
        get:
            tags:
                - control-plane
            summary: list-databases control-plane
            description: Lists all databases in the cluster.
            operationId: control-plane#list-databases
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/ControlPlaneDatabaseResponseAbbreviatedCollection'
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ControlPlaneListDatabasesClusterNotInitializedResponseBody'
            schemes:
                - http
        post:
            tags:
                - control-plane
            summary: create-database control-plane
            description: Creates a new database in the cluster.
            operationId: control-plane#create-database
            parameters:
                - name: Create-DatabaseRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/CreateDatabaseRequest'
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/ControlPlaneCreateDatabaseResponseBody'
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/ControlPlaneCreateDatabaseInvalidInputResponseBody'
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ControlPlaneCreateDatabaseClusterNotInitializedResponseBody'
            schemes:
                - http
    /databases/{database_id}:
        get:
            tags:
                - control-plane
            summary: inspect-database control-plane
            description: Returns information about a particular database in the cluster.
            operationId: control-plane#inspect-database
            parameters:
                - name: database_id
                  in: path
                  description: ID of the database to inspect.
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/ControlPlaneInspectDatabaseResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/ControlPlaneInspectDatabaseNotFoundResponseBody'
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ControlPlaneInspectDatabaseClusterNotInitializedResponseBody'
            schemes:
                - http
        post:
            tags:
                - control-plane
            summary: update-database control-plane
            description: Updates a database with the given specification.
            operationId: control-plane#update-database
            parameters:
                - name: force_update
                  in: query
                  description: Force update the database even if the spec is the same.
                  required: false
                  type: boolean
                - name: database_id
                  in: path
                  description: ID of the database to update.
                  required: true
                  type: string
                - name: Update-DatabaseRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/UpdateDatabaseRequest'
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/ControlPlaneUpdateDatabaseResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/ControlPlaneUpdateDatabaseNotFoundResponseBody'
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ControlPlaneUpdateDatabaseClusterNotInitializedResponseBody'
            schemes:
                - http
        delete:
            tags:
                - control-plane
            summary: delete-database control-plane
            description: Deletes a database from the cluster.
            operationId: control-plane#delete-database
            parameters:
                - name: database_id
                  in: path
                  description: ID of the database to delete.
                  required: true
                  type: string
            responses:
                "204":
                    description: No Content response.
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/ControlPlaneDeleteDatabaseNotFoundResponseBody'
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ControlPlaneDeleteDatabaseClusterNotInitializedResponseBody'
            schemes:
                - http
    /hosts:
        get:
            tags:
                - control-plane
            summary: list-hosts control-plane
            description: Lists all hosts within the cluster.
            operationId: control-plane#list-hosts
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/Host'
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ControlPlaneListHostsClusterNotInitializedResponseBody'
            schemes:
                - http
    /hosts/{host_id}:
        get:
            tags:
                - control-plane
            summary: inspect-host control-plane
            description: Returns information about a particular host in the cluster.
            operationId: control-plane#inspect-host
            parameters:
                - name: host_id
                  in: path
                  description: ID of the host to inspect.
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Host'
                        required:
                            - id
                            - orchestrator
                            - hostname
                            - ipv4_address
                            - cpus
                            - memory
                            - status
                            - default_pgedge_version
                            - supported_pgedge_versions
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/ControlPlaneInspectHostNotFoundResponseBody'
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ControlPlaneInspectHostClusterNotInitializedResponseBody'
            schemes:
                - http
        delete:
            tags:
                - control-plane
            summary: remove-host control-plane
            description: Removes a host from the cluster.
            operationId: control-plane#remove-host
            parameters:
                - name: host_id
                  in: path
                  description: ID of the host to remove.
                  required: true
                  type: string
            responses:
                "204":
                    description: No Content response.
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/ControlPlaneRemoveHostNotFoundResponseBody'
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ControlPlaneRemoveHostClusterNotInitializedResponseBody'
            schemes:
                - http
...
definitions:
...
    Cluster:
        title: Cluster
        type: object
        properties:
            hosts:
                type: array
                items:
                    $ref: '#/definitions/Host'
                description: All of the hosts in the cluster.
                example:
                    - cohort:
                        cohort_id: pps1n11hqijn9rbee4cjil453
                        control_available: true
                        member_id: lah4bsznw6kc0hp7biylmmmll
                        type: swarm
                      cpus: 4
                      default_pgedge_version:
                        postgres_version: "16"
                        spock_version: "4"
                      hostname: i-0123456789abcdef.ec2.internal
                      id: de3b1388-1f0c-42f1-a86c-59ab72f255ec
                      ipv4_address: 10.24.34.0
                      memory: 16GiB
                      orchestrator: swarm
                      status:
                        components:
                            Eum sequi enim.:
                                details:
                                    alarms:
                                        - '3: NOSPACE'
                                error: failed to connect to etcd
                                healthy: false
                            Libero repellendus qui nobis.:
                                details:
                                    alarms:
                                        - '3: NOSPACE'
                                error: failed to connect to etcd
                                healthy: false
                        state: available
                        updated_at: "2021-07-01T12:34:56Z"
                      supported_pgedge_versions:
                        - postgres_version: "16"
                          spock_version: "4"
                        - postgres_version: "16"
                          spock_version: "4"
                        - postgres_version: "16"
                          spock_version: "4"
                    - cohort:
                        cohort_id: pps1n11hqijn9rbee4cjil453
                        control_available: true
                        member_id: lah4bsznw6kc0hp7biylmmmll
                        type: swarm
                      cpus: 4
                      default_pgedge_version:
                        postgres_version: "16"
                        spock_version: "4"
                      hostname: i-0123456789abcdef.ec2.internal
                      id: de3b1388-1f0c-42f1-a86c-59ab72f255ec
                      ipv4_address: 10.24.34.0
                      memory: 16GiB
                      orchestrator: swarm
                      status:
                        components:
                            Eum sequi enim.:
                                details:
                                    alarms:
                                        - '3: NOSPACE'
                                error: failed to connect to etcd
                                healthy: false
                            Libero repellendus qui nobis.:
                                details:
                                    alarms:
                                        - '3: NOSPACE'
                                error: failed to connect to etcd
                                healthy: false
                        state: available
                        updated_at: "2021-07-01T12:34:56Z"
                      supported_pgedge_versions:
                        - postgres_version: "16"
                          spock_version: "4"
                        - postgres_version: "16"
                          spock_version: "4"
                        - postgres_version: "16"
                          spock_version: "4"
            id:
                type: string
                description: Unique identifier for the cluster.
                example: a67cbb36-c3c3-49c9-8aac-f4a0438a883d
                format: uuid
            status:
                $ref: '#/definitions/ClusterStatus'
            tenant_id:
                type: string
                description: Unique identifier for the cluster's owner.
                example: 8210ec10-2dca-406c-ac4a-0661d2189954
                format: uuid
        example:
            hosts:
                - cohort:
                    cohort_id: pps1n11hqijn9rbee4cjil453
                    control_available: true
                    member_id: lah4bsznw6kc0hp7biylmmmll
                    type: swarm
                  cpus: 4
                  default_pgedge_version:
                    postgres_version: "16"
                    spock_version: "4"
                  hostname: i-0123456789abcdef.ec2.internal
                  id: de3b1388-1f0c-42f1-a86c-59ab72f255ec
                  ipv4_address: 10.24.34.0
                  memory: 16GiB
                  orchestrator: swarm
                  status:
                    components:
                        Eum sequi enim.:
                            details:
                                alarms:
                                    - '3: NOSPACE'
                            error: failed to connect to etcd
                            healthy: false
                        Libero repellendus qui nobis.:
                            details:
                                alarms:
                                    - '3: NOSPACE'
                            error: failed to connect to etcd
                            healthy: false
                    state: available
                    updated_at: "2021-07-01T12:34:56Z"
                  supported_pgedge_versions:
                    - postgres_version: "16"
                      spock_version: "4"
                    - postgres_version: "16"
                      spock_version: "4"
                    - postgres_version: "16"
                      spock_version: "4"
                - cohort:
                    cohort_id: pps1n11hqijn9rbee4cjil453
                    control_available: true
                    member_id: lah4bsznw6kc0hp7biylmmmll
                    type: swarm
                  cpus: 4
                  default_pgedge_version:
                    postgres_version: "16"
                    spock_version: "4"
                  hostname: i-0123456789abcdef.ec2.internal
                  id: de3b1388-1f0c-42f1-a86c-59ab72f255ec
                  ipv4_address: 10.24.34.0
                  memory: 16GiB
                  orchestrator: swarm
                  status:
                    components:
                        Eum sequi enim.:
                            details:
                                alarms:
                                    - '3: NOSPACE'
                            error: failed to connect to etcd
                            healthy: false
                        Libero repellendus qui nobis.:
                            details:
                                alarms:
                                    - '3: NOSPACE'
                            error: failed to connect to etcd
                            healthy: false
                    state: available
                    updated_at: "2021-07-01T12:34:56Z"
                  supported_pgedge_versions:
                    - postgres_version: "16"
                      spock_version: "4"
                    - postgres_version: "16"
                      spock_version: "4"
                    - postgres_version: "16"
                      spock_version: "4"
                - cohort:
                    cohort_id: pps1n11hqijn9rbee4cjil453
                    control_available: true
                    member_id: lah4bsznw6kc0hp7biylmmmll
                    type: swarm
                  cpus: 4
                  default_pgedge_version:
                    postgres_version: "16"
                    spock_version: "4"
                  hostname: i-0123456789abcdef.ec2.internal
                  id: de3b1388-1f0c-42f1-a86c-59ab72f255ec
                  ipv4_address: 10.24.34.0
                  memory: 16GiB
                  orchestrator: swarm
                  status:
                    components:
                        Eum sequi enim.:
                            details:
                                alarms:
                                    - '3: NOSPACE'
                            error: failed to connect to etcd
                            healthy: false
                        Libero repellendus qui nobis.:
                            details:
                                alarms:
                                    - '3: NOSPACE'
                            error: failed to connect to etcd
                            healthy: false
                    state: available
                    updated_at: "2021-07-01T12:34:56Z"
                  supported_pgedge_versions:
                    - postgres_version: "16"
                      spock_version: "4"
                    - postgres_version: "16"
                      spock_version: "4"
                    - postgres_version: "16"
                      spock_version: "4"
            id: a67cbb36-c3c3-49c9-8aac-f4a0438a883d
            status:
                state: error
            tenant_id: 8210ec10-2dca-406c-ac4a-0661d2189954
        required:
            - id
            - tenant_id
            - status
            - hosts
    ClusterJoinToken:
        title: ClusterJoinToken
        type: object
        properties:
            server_url:
                type: string
                description: Existing server to join
                example: http://192.168.1.1:3000
                format: uri
            token:
                type: string
                description: Token to join an existing cluster.
                example: PGEDGE-dd440afcf5de20ef8e8cf54f6cb9f125fd55f90e64faa94b906130b31235e730-41e975f41d7ea61058f2fe2572cb52dd
        example:
            server_url: http://192.168.1.1:3000
            token: PGEDGE-dd440afcf5de20ef8e8cf54f6cb9f125fd55f90e64faa94b906130b31235e730-41e975f41d7ea61058f2fe2572cb52dd
        required:
            - token
            - server_url
    ClusterStatus:
        title: ClusterStatus
        type: object
        properties:
            state:
                type: string
                description: The current state of the cluster.
                example: available
                enum:
                    - available
                    - error
        example:
            state: available
        required:
            - state
...
    ControlPlaneInitClusterClusterAlreadyInitializedResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: false
        description: init-cluster_cluster_already_initialized_response_body result type (default view)
        example:
            fault: true
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: false
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    ControlPlaneInspectClusterClusterNotInitializedResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: true
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: true
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: false
        description: inspect-cluster_cluster_not_initialized_response_body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: false
            timeout: false
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    ControlPlaneInspectClusterNotFoundResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: true
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: inspect-cluster_not_found_response_body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
...
    ControlPlaneJoinClusterClusterAlreadyInitializedResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: true
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: true
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: join-cluster_cluster_already_initialized_response_body result type (default view)
        example:
            fault: true
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: false
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
...